requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.6
django-environ==0.4.5
django-filter==2.4.0
django-mail-templated==2.6.5
django-safedelete==1.0.0
django-seed==0.2.2
djangorestframework==3.12.4
drf-spectacular==0.17.3
Faker==8.11.0
inflection==0.5.1
jsonschema==3.2.0
psycopg2-binary==2.9.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
